! Time offset wasn't casting properly.
git-svn-id: https://devel.simplemachines.org/code/smf/trunk@6585 a96bbafc-d752-43aa-a6db-eed64a5f2cf9

+++ Sources/Profile-Modify.php
@@ -741,7 +741,7 @@ function loadProfileFields($force_reload = false)
 			'),
 			'input_validate' => create_function('&$value', '
 				// Validate the time_offset...
-				$value = strtr($value, \',\', \'.\');
+				$value = (float) strtr($value, \',\', \'.\');
 
 				if ($value < -23.5 || $value > 23.5)
 					return \'bad_offset\';
